requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
gspread
google-auth
Jinja2
openai
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=0.8.11
numpy>=1.24
sentence-transformers>=2.7.0
transformers>=4.41.0

